Separate convert module from utils
